Added test for admins to read all current matters
And fixed new eslint rule prohibiting default exports by default.
