+++ packages/api/tests/utils/dbHelpers.ts
@@ -3,6 +3,8 @@ import { Pool } from 'pg';
 import { afterAll } from '@jest/globals';
 import { postgresUrl } from '../../src/postgresUrl';
 
+export { createMatter } from './db/createMatter';
+
 const pools: any = {};
 
 // Make sure we release those pgPools so that our tests exit!
@@ -132,59 +134,6 @@ export const createFlashcard = async (client: any) => {
   );
 
   return rows[0];
-};
-
-interface CreateMatterArgs {
-  client: any;
-  primaryContactId?: string;
-  matterTemplateId?: string;
-}
-
-export const createMatter = async ({
-  client,
-  primaryContactId,
-  matterTemplateId,
-}: CreateMatterArgs) => {
-  const uuid = faker.random.uuid();
-
-  let matterTemplateIdForInsertingMatter;
-
-  if (matterTemplateId) {
-    matterTemplateIdForInsertingMatter = matterTemplateId;
-  } else {
-    const { rows: matterTemplateRows } = await client.query(
-      'INSERT INTO matter_template (name, javascript_module) '+
-          'VALUES ($1, $2) RETURNING (id)',
-      [`delete-your-data-${uuid}`, `deleteYourData-${uuid}`]
-    );
-    matterTemplateIdForInsertingMatter = matterTemplateRows[0].id;
-  }
-
-  let primaryContactIdForInsertingMatter;
-
-  if (primaryContactId) {
-    primaryContactIdForInsertingMatter = primaryContactId;
-  } else {
-    const { rows: primaryContactRows } = await client.query(
-      'INSERT INTO person (email, role, sub) ' +
-          'VALUES ($1, $2, $3) RETURNING (id)',
-      [`example-contact-${uuid}@neonlaw.com`, 'portal', `portal-${uuid}`]
-    );
-    primaryContactIdForInsertingMatter = primaryContactRows[0].id;
-  }
-
-  const { rows } = await client.query(
-    'INSERT INTO matter (name, primary_contact_id, '+
-          'matter_template_id) VALUES ($1, $2, $3) '+
-          'RETURNING (id, primary_contact_id, matter_template_id)',
-    [
-      `Random matter ${uuid}`,
-      primaryContactIdForInsertingMatter,
-      matterTemplateIdForInsertingMatter
-    ]
-  );
-
-  return rows[0].row;
 };
 
 export const createMatterTemplate = async (
