+++ packages/delete-your-data/src/layouts/postLayout.tsx
@@ -147,6 +147,7 @@ const PostLayout: React.FC<{
   );
 };
 
+/* eslint-disable-next-line import/no-default-export */
 export default PostLayout;
 
 export const pageQuery = graphql`
