+++ packages/delete-your-data/src/pages/index.tsx
@@ -1,4 +1,4 @@
-import FAQs from '../components/index/faqs';
+import { FAQs } from '../components/index/faqs';
 import { Hero } from '../components/hero';
 import IndexLayout from '../layouts';
 import React from 'react';
