+++ packages/interface/src/components/homepage/experience.tsx
@@ -68,5 +68,3 @@ export const Experience = () => {
     </Section>
   );
 };
-
-export default Experience;
