+++ packages/interface/src/layouts/mdxLayout.tsx
@@ -101,6 +101,7 @@ const MdxLayout: React.FC<{
   );
 };
 
+/* eslint-disable-next-line import/no-default-export */
 export default MdxLayout;
 
 export const pageQuery = graphql`
