+++ packages/justice-for-rickie-slaughter/src/layouts/mdxLayout.tsx
@@ -90,6 +90,7 @@ const MdxLayout: React.FC<{
   );
 };
 
+/* eslint-disable-next-line import/no-default-export */
 export default MdxLayout;
 
 export const pageQuery = graphql`
