+++ packages/law-job-resources/src/layouts/mdxLayout.tsx
@@ -70,6 +70,7 @@ const MdxLayout: React.FC<{
   );
 };
 
+/* eslint-disable-next-line import/no-default-export */
 export default MdxLayout;
 
 export const pageQuery = graphql`
