+++ packages/shared-ui/src/layouts/baseLayout.tsx
@@ -3,7 +3,7 @@ import React, { ReactChildren, useEffect } from 'react';
 import { AuthenticationProvider } from '../utils/authenticationContext';
 import { Background } from '../components/background';
 import { BaseStyles } from '../styles/baseStyles';
-import ShortcutsModal from '../components/shortcuts-modal';
+import { ShortcutsModal } from '../components/shortcutsModal';
 import { handleFirstTab } from '../utils/accessibility';
 
 export const BaseLayout: React.FC<{
